pyproject.toml
[project]
name = "consents-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-bitfield>=2.2.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.15.2",
    "drf-nested-routers>=0.94.1",
    "drf-yasg>=1.21.8",
    "orjson>=3.10.16",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
]
